Refactor project structure
Also removed unused components
