--- src/components/customSelect.tsx
@@ -1,44 +1,0 @@
-import Select, { SelectChangeEvent } from '@mui/material/Select';
-import FormControl from '@mui/material/FormControl';
-import InputLabel from '@mui/material/InputLabel';
-import MenuItem from '@mui/material/MenuItem';
-import Box from '@mui/material/Box';
-
-type SelectProps = {
-    label: string;
-    options: string[];
-    required: boolean;
-    value: string;
-    onChange: Function;
-};
-
-const CustomSelect = ({
-    label,
-    options,
-    required,
-    value,
-    onChange,
-}: SelectProps) => {
-    const handleChange = (event: SelectChangeEvent) => {
-        onChange(event.target.value as string);
-    };
-
-    return (
-        <Box>
-            <FormControl fullWidth required={required}>
-                <InputLabel>{label}</InputLabel>
-                <Select value={value} label={label} onChange={handleChange}>
-                    {options.map((option: string) => {
-                        return (
-                            <MenuItem key={option} value={option}>
-                                {option}
-                            </MenuItem>
-                        );
-                    })}
-                </Select>
-            </FormControl>
-        </Box>
-    );
-};
-
-export default CustomSelect;
