+++ src/layouts/MainLayout.tsx
@@ -4,12 +4,12 @@ import { styled, useTheme, Theme, CSSObject } from '@mui/material/styles';
 import ViewListRoundedIcon from '@mui/icons-material/ViewListRounded';
 import { capitalizeEachWord } from '../utils/capitalizeEachWord.tsx';
 import ManageAccountsIcon from '@mui/icons-material/ManageAccounts';
+import { AccountMenu } from '../components/menus/AccountMenu.tsx';
 import AccountCircleIcon from '@mui/icons-material/AccountCircle';
 import ChevronRightIcon from '@mui/icons-material/ChevronRight';
 import ChevronLeftIcon from '@mui/icons-material/ChevronLeft';
 import HomeRoundedIcon from '@mui/icons-material/HomeRounded';
 import ListItemButton from '@mui/material/ListItemButton';
-import AccountMenu from '../components/AccountMenu.tsx';
 import DatasetIcon from '@mui/icons-material/Dataset';
 import ListItemIcon from '@mui/material/ListItemIcon';
 import ListItemText from '@mui/material/ListItemText';
