+++ src/renderers/renderBidID.tsx
