+++ src/renderers/renderCurrency.tsx
