+++ src/renderers/renderPercentage.tsx
